Show hero image even if no js
By setting it as a background image for the home section, and removing
it through javascript (i.e. if javascript is enabled, use parallax
feature instead).
I am not going for a JS-less site (for now). I like the adaptive menu.
Maybe I'll want the masonry or other features too.